+++ R/stats.R
@@ -168,6 +168,8 @@ dgr.array <- function(graph, cmode, undirected, self, valued) {
 #' @param valued Logical scalar. When \code{FALSE}, values of \code{graph} are set to one.
 #' @param normalized Logical scalar. When \code{TRUE}, the exposure will be between zero
 #' and one (see details).
+#' @param groupvar Character scalar. Name of a vertex static attribute in the diffnet
+#' object (see details).
 #' @param ... Further arguments passed to \code{\link{struct_equiv}} (only used when
 #' \code{alt.graph="se"}).
 #' @details
@@ -194,6 +196,18 @@ dgr.array <- function(graph, cmode, undirected, self, valued) {
 #' graph. Notice that when using a valued graph the option \code{valued} should
 #' be equal to \code{TRUE}, this check is run automatically when running the
 #' model using structural equivalence.
+#'
+#' \bold{An important remark} is that when calculating \bold{structural equivalence} the
+#' function \bold{assumes that this is to be done to the entire graph} regardless of
+#' disconnected communities (as in the case of the medical innovations
+#' data set). Hence, structural equivalence for individuals for two different
+#' communites may not be zero. If the user wants to calculate structural
+#' equivalence separately by community, he should create different diffnet
+#' objects and do so (see example below). Alternatively, for the case of
+#' diffnet objects, by using the option \code{groupvar}, the user can provide
+#' the function with the name of a grouping variable--which should one in the
+#' set of static vertex attributes--so that the algorithm is done by group
+#' (or community) instead of in an aggregated way.
 #'
 #' If the user does not specifies a particular weighting attribute in \code{attrs},
 #' the function sets this as a matrix of ones. Otherwise the function will return
@@ -224,13 +238,19 @@ dgr.array <- function(graph, cmode, undirected, self, valued) {
 #' graph <- rdiffnet(100, 10)
 #'
 #' SE <- lapply(struct_equiv(graph), "[[", "SE")
-#' SE <- lapply(SE, function(x) 1/(x + 1e-15))
+#' SE <- lapply(SE, function(x) {
+#'    x <- 1/x
+#'    x[!is.finite(x)] <- 0
+#'    x
+#' })
 #'
 #' # Recall setting valued equal to TRUE!
 #' expo_se <- exposure(graph, alt.graph=SE , valued=TRUE)
 #'
-#' # These three lines are equivalent to
-#' expo_se2 <- exposure(graph, alt.graph="se")
+#' # These three lines are equivalent to:
+#' expo_se2 <- exposure(graph, alt.graph="se", valued=TRUE)
+#' # Notice that we are setting valued=TRUE, but this is not necesary since when
+#' # alt.graph = "se" the function checks this to be setted equal to TRUE
 #'
 #' # Weighted Exposure using degree --------------------------------------------
 #' eDE <- exposure(graph, attrs=dgr(graph))
@@ -242,6 +262,62 @@ dgr.array <- function(graph, cmode, undirected, self, valued) {
 #' # Comparing using incomming edges -------------------------------------------
 #' eIN <- exposure(graph, outgoing=FALSE)
 #'
+#' # Structral equivalence for different communities ---------------------------
+#' data(brfarmersDiffNet)
+#'
+#' # METHOD 1: Using the c.diffnet method:
+#'
+#' # Creating subsets by village
+#' villages <- unique(brfarmersDiffNet[["village"]])
+#'
+#' diffnet <- brfarmersDiffNet[brfarmersDiffNet[["village"]] == villages[1]]
+#' diffnet[["expo_se"]] <- exposure(diffnet, alt.graph="se", valued=TRUE)
+#'
+#' for (v in villages[-1]) {
+#'    diffnet_v <- brfarmersDiffNet[brfarmersDiffNet[["village"]] == v]
+#'    diffnet_v[["expo_se"]] <- exposure(diffnet_v, alt.graph="se", valued=TRUE)
+#'    diffnet <- c(diffnet, diffnet_v)
+#' }
+#'
+#' # We can set the original order (just in case) of the data
+#' diffnet <- diffnet[brfarmersDiffNet$meta$ids]
+#' diffnet
+#'
+#' # Checking everything is equal
+#' test <- summary(brfarmersDiffNet, no.print=TRUE) ==
+#'    summary(diffnet, no.print=TRUE)
+#'
+#' stopifnot(all(test))
+#'
+#' # METHOD 2: Using the 'groupvar' argument
+#' # Further, we can compare this with using the groupvar
+#' diffnet[["expo_se2"]] <- exposure(diffnet, alt.graph="se",
+#'    groupvar="village", valued=TRUE)
+#'
+#' # These should be equivalent
+#' test <- diffnet[["expo_se", as.df=TRUE]] == diffnet[["expo_se2", as.df=TRUE]]
+#' stopifnot(all(test))
+#'
+#' # METHOD 3: Computing exposure, rbind and then adding it to the diffnet object
+#' expo_se3 <- NULL
+#' for (v in unique(villages))
+#'    expo_se3 <- rbind(
+#'      expo_se3,
+#'      exposure(
+#'        diffnet[diffnet[["village"]] == v],
+#'        alt.graph = "se", valued=TRUE
+#'      ))
+#'
+#' # Just to make sure, we sort the rows
+#' expo_se3 <- expo_se3[diffnet$meta$ids,]
+#'
+#' diffnet[["expo_se3"]] <- expo_se3
+#'
+#' test <- diffnet[["expo_se", as.df=TRUE]] == diffnet[["expo_se3", as.df=TRUE]]
+#' stopifnot(all(test))
+#'
+#'
+#'
 #' @family statistics
 #' @keywords univar
 #' @return A matrix of size \eqn{n\times T}{n * T} with exposure for each node.
@@ -249,7 +325,41 @@ dgr.array <- function(graph, cmode, undirected, self, valued) {
 #' @author Vega Yon, Dyal, Hayes & Valente
 exposure <- function(graph, cumadopt, attrs = NULL, alt.graph=NULL,
                      outgoing=getOption("diffnet.outgoing", TRUE),
-                     valued=getOption("diffnet.valued", FALSE), normalized=TRUE, ...) {
+                     valued=getOption("diffnet.valued", FALSE), normalized=TRUE,
+                     groupvar=NULL, ...) {
+
+  # Checking if there exists a group var
+  if (length(groupvar)) {
+    if (inherits(graph, "diffnet")) {
+
+      # Analyzing groups
+      groups <- graph[[groupvar]]
+      G      <- unique(groups)
+      out    <- NULL
+      for (g in G) {
+        # Getting indexes
+        index <- which(groups==g)
+
+        if (length(alt.graph) && is.list(alt.graph))
+          alt.g <- lapply(alt.graph, function(x) x[index,index,drop=FALSE])
+        else if (length(alt.graph) && alt.graph == "se")
+          alt.g <- alt.graph
+        else alt.g <- NULL
+
+        # Computing exposure
+        out <- rbind(out,
+          exposure(graph=graph[index], attrs=attrs[index], alt.graph=alt.g,
+                   outgoing = outgoing, valued = valued,normalized = normalized,
+                   ...))
+      }
+
+      # Sorting the data and returning
+      out <- out[graph$meta$ids,]
+      return(out)
+    } else {
+      stop("-groupvar- is currently only supported for diffnet objects.")
+    }
+  }
 
   # Checking diffnet attributes
   if (length(attrs) == 1 && inherits(attrs, "character")) {
@@ -283,9 +393,14 @@ exposure <- function(graph, cumadopt, attrs = NULL, alt.graph=NULL,
     graph <- if (inherits(alt.graph, "character")) {
       if (alt.graph != "se") stop("Only character -alt.graph- value allowed is \"se\".")
 
-      # Computing structural equivalence
       se <- lapply(struct_equiv(graph,...), "[[", "SE")
-      se <- lapply(se, function(x) 1/(x + 1e-15))
+      se <- lapply(se, function(x) {
+        x <- 1/x
+        x[!is.finite(x)] <- 0
+        x
+        })
+      # }
+
 
       # Changing valued
       if (!valued) {