+++ data-raw/brfarmers.R
@@ -3,4 +3,5 @@ library(foreign)
 
 # Preparing the data -----------------------------------------------------------
 brfarmers <- read.dta("data-raw/brfarmers.dta")
-save(brfarmers,file =  "data/brfarmers.rdata")
+save(brfarmers,file =  "data/brfarmers.rdata",
+     compress = "xz")