+++ data-raw/medInnovations.R
@@ -4,4 +4,5 @@ library(foreign)
 # Preparing the data -----------------------------------------------------------
 medInnovations <- read.dta("data-raw/mi_v2.dta")
 
-save(medInnovations, file="data/medInnovations.rdata")
+save(medInnovations, file="data/medInnovations.rdata",
+     compress = "xz")